feat: Refactor include directory name to be consistent with the new package name.
