+++ include/mpc_rbt_simulator/RobotConfig.hpp
